WebSockets/tests/submissions/Opera/interfaces/WebSocket/events/013.html

<!doctype html>
<title>WebSockets: setting event handlers to {}</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js></script>
<div id=log></div>
<script>
var events = ['onclose', 'onopen', 'onerror', 'onmessage'];
for (var i = 0; i < events.length; ++i) {
  async_test(function(t) {
    var ws = new WebSocket(SCHEME_AND_DOMAIN+'/');
    assert_throws(new TypeError(), function(){
      ws[events[i]] = {};
    });
  })
};
</script>
